Fix task modification bug
Mainly caused by wrong datatypes in database

+++ app/database/migrations/2014_06_26_131131_create_modified_date_tasks_table.php
@@ -16,7 +16,7 @@ class CreateModifiedDateTasksTable extends Migration {
 		{
 			$table->increments('id');
 
-			$table->integer('asana_task_id');
+			$table->bigInteger('asana_task_id');
 			$table->index('asana_task_id');
 
 			$table->date('modified_date');
