+++ app/database/migrations/2014_06_27_132600_create_modified_name_tasks_table.php
@@ -16,7 +16,7 @@ class CreateModifiedNameTasksTable extends Migration {
 		{
 			$table->increments('id');
 
-			$table->integer('asana_task_id');
+			$table->bigInteger('asana_task_id');
 			$table->index('asana_task_id');
 
 			$table->string('modified_title');
